+++ src/main/kotlin/blackjack/domain/card/CardDeck.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import java.util.LinkedList
 import java.util.Queue
