+++ src/main/kotlin/blackjack/domain/card/Cards.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 @JvmInline
 value class Cards(
