+++ src/main/kotlin/blackjack/domain/gamestate/GameState.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.gamestate
 
-import blackjack.domain.Card
+import blackjack.domain.card.Card
 
 interface GameState {
