+++ src/test/kotlin/blackjack/domain/gamestate/HitTest.kt
@@ -1,12 +1,12 @@
 package blackjack.domain.gamestate
 
-import blackjack.domain.CardTest.Companion.SPADE_ACE
-import blackjack.domain.CardTest.Companion.SPADE_JACK
-import blackjack.domain.CardTest.Companion.SPADE_KING
-import blackjack.domain.CardTest.Companion.SPADE_QUEEN
-import blackjack.domain.CardTest.Companion.SPADE_THREE
-import blackjack.domain.CardTest.Companion.SPADE_TWO
-import blackjack.domain.Cards
+import blackjack.domain.card.CardTest.Companion.SPADE_ACE
+import blackjack.domain.card.CardTest.Companion.SPADE_JACK
+import blackjack.domain.card.CardTest.Companion.SPADE_KING
+import blackjack.domain.card.CardTest.Companion.SPADE_QUEEN
+import blackjack.domain.card.CardTest.Companion.SPADE_THREE
+import blackjack.domain.card.CardTest.Companion.SPADE_TWO
+import blackjack.domain.card.Cards
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldContainAll
