+++ src/test/kotlin/blackjack/domain/gamestate/StayTest.kt
@@ -1,9 +1,9 @@
 package blackjack.domain.gamestate
 
-import blackjack.domain.CardTest.Companion.SPADE_ACE
-import blackjack.domain.CardTest.Companion.SPADE_JACK
-import blackjack.domain.CardTest.Companion.SPADE_KING
-import blackjack.domain.Cards
+import blackjack.domain.card.CardTest.Companion.SPADE_ACE
+import blackjack.domain.card.CardTest.Companion.SPADE_JACK
+import blackjack.domain.card.CardTest.Companion.SPADE_KING
+import blackjack.domain.card.Cards
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
